refactor: move all batch and product contents from marketplace to warehouse
