+++ src/main/java/br/com/mercadolivre/projetointegrador/warehouse/controller/BatchController.java
@@ -1,8 +1,8 @@
-package br.com.mercadolivre.projetointegrador.marketplace.controller;
+package br.com.mercadolivre.projetointegrador.warehouse.controller;
 
-import br.com.mercadolivre.projetointegrador.marketplace.exception.NotFoundException;
-import br.com.mercadolivre.projetointegrador.marketplace.model.Batch;
-import br.com.mercadolivre.projetointegrador.marketplace.service.BatchService;
+import br.com.mercadolivre.projetointegrador.warehouse.exception.db.NotFoundException;
+import br.com.mercadolivre.projetointegrador.warehouse.model.Batch;
+import br.com.mercadolivre.projetointegrador.warehouse.service.BatchService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
