+++ src/main/java/br/com/mercadolivre/projetointegrador/warehouse/controller/PingController.java
@@ -1,4 +1,4 @@
-package br.com.mercadolivre.projetointegrador.marketplace.controller;
+package br.com.mercadolivre.projetointegrador.warehouse.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
