+++ src/main/java/br/com/mercadolivre/projetointegrador/warehouse/dto/response/CreatedBatchDTO.java
@@ -1,7 +1,7 @@
 package br.com.mercadolivre.projetointegrador.warehouse.dto.response;
 
-import br.com.mercadolivre.projetointegrador.marketplace.model.Batch;
-import br.com.mercadolivre.projetointegrador.marketplace.model.Product;
+import br.com.mercadolivre.projetointegrador.warehouse.model.Batch;
+import br.com.mercadolivre.projetointegrador.warehouse.model.Product;
 import lombok.*;
 
 import java.math.BigDecimal;
