+++ src/main/java/br/com/mercadolivre/projetointegrador/warehouse/exception/ErrorDTO.java
@@ -1,4 +1,4 @@
-package br.com.mercadolivre.projetointegrador.marketplace.exception;
+package br.com.mercadolivre.projetointegrador.warehouse.exception;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
