+++ src/main/java/br/com/mercadolivre/projetointegrador/warehouse/exception/db/ProductAlreadyExists.java
@@ -1,4 +1,4 @@
-package br.com.mercadolivre.projetointegrador.marketplace.exception;
+package br.com.mercadolivre.projetointegrador.warehouse.exception.db;
 
 public class ProductAlreadyExists extends Exception {
