+++ src/main/java/br/com/mercadolivre/projetointegrador/warehouse/model/Batch.java
@@ -1,5 +1,6 @@
-package br.com.mercadolivre.projetointegrador.marketplace.model;
+package br.com.mercadolivre.projetointegrador.warehouse.model;
 
+import br.com.mercadolivre.projetointegrador.warehouse.model.Product;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
