+++ src/main/java/br/com/mercadolivre/projetointegrador/warehouse/model/Product.java
@@ -1,6 +1,6 @@
-package br.com.mercadolivre.projetointegrador.marketplace.model;
+package br.com.mercadolivre.projetointegrador.warehouse.model;
 
-import br.com.mercadolivre.projetointegrador.marketplace.enums.CategoryEnum;
+import br.com.mercadolivre.projetointegrador.warehouse.enums.CategoryEnum;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
