+++ src/main/java/br/com/mercadolivre/projetointegrador/warehouse/service/ProductService.java
@@ -1,9 +1,9 @@
-package br.com.mercadolivre.projetointegrador.marketplace.service;
+package br.com.mercadolivre.projetointegrador.warehouse.service;
 
-import br.com.mercadolivre.projetointegrador.marketplace.exception.NotFoundException;
-import br.com.mercadolivre.projetointegrador.marketplace.exception.ProductAlreadyExists;
-import br.com.mercadolivre.projetointegrador.marketplace.model.Product;
-import br.com.mercadolivre.projetointegrador.marketplace.repository.ProductRepository;
+import br.com.mercadolivre.projetointegrador.warehouse.exception.db.NotFoundException;
+import br.com.mercadolivre.projetointegrador.warehouse.exception.db.ProductAlreadyExists;
+import br.com.mercadolivre.projetointegrador.warehouse.model.Product;
+import br.com.mercadolivre.projetointegrador.warehouse.repository.ProductRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
