+++ src/test/java/br/com/mercadolivre/projetointegrador/integration/controller/SectionControllerTests.java
@@ -1,6 +1,6 @@
 package br.com.mercadolivre.projetointegrador.integration.controller;
 
-import br.com.mercadolivre.projetointegrador.marketplace.enums.CategoryEnum;
+import br.com.mercadolivre.projetointegrador.warehouse.enums.CategoryEnum;
 import br.com.mercadolivre.projetointegrador.test_utils.IntegrationTestUtils;
 import br.com.mercadolivre.projetointegrador.warehouse.dto.request.CreateSectionPayloadDTO;
 import br.com.mercadolivre.projetointegrador.warehouse.model.Section;
