+++ src/test/java/br/com/mercadolivre/projetointegrador/unit/service/BatchServiceTests.java
@@ -1,13 +1,13 @@
 package br.com.mercadolivre.projetointegrador.unit.service;
 
-import br.com.mercadolivre.projetointegrador.marketplace.exception.NotFoundException;
-import br.com.mercadolivre.projetointegrador.marketplace.exception.ProductAlreadyExists;
-import br.com.mercadolivre.projetointegrador.marketplace.model.Batch;
-import br.com.mercadolivre.projetointegrador.marketplace.model.Product;
-import br.com.mercadolivre.projetointegrador.marketplace.repository.BatchRepository;
-import br.com.mercadolivre.projetointegrador.marketplace.repository.ProductRepository;
-import br.com.mercadolivre.projetointegrador.marketplace.service.BatchService;
-import br.com.mercadolivre.projetointegrador.marketplace.service.ProductService;
+import br.com.mercadolivre.projetointegrador.warehouse.exception.db.NotFoundException;
+import br.com.mercadolivre.projetointegrador.warehouse.exception.db.ProductAlreadyExists;
+import br.com.mercadolivre.projetointegrador.warehouse.model.Batch;
+import br.com.mercadolivre.projetointegrador.warehouse.model.Product;
+import br.com.mercadolivre.projetointegrador.warehouse.repository.BatchRepository;
+import br.com.mercadolivre.projetointegrador.warehouse.repository.ProductRepository;
+import br.com.mercadolivre.projetointegrador.warehouse.service.BatchService;
+import br.com.mercadolivre.projetointegrador.warehouse.service.ProductService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
